requirements.txt
langchain-sqlserver==0.1.1
langchain-huggingface
sentence-transformers
langchain-community
faiss-cpu